bring many files into browser.js, routes and style left alone
