--- client.js
@@ -1,7 +1,0 @@
-const authentication = require('dogstack-agents/client')
-
-export default {
-  services: [
-    authentication
-  ]
-}
