--- intl.js
@@ -1,12 +1,0 @@
-import { addLocaleData } from 'react-intl'
-import en from 'react-intl/locale-data/en'
-
-addLocaleData([...en])
-
-const messagesByLocale = {
-  'en': require('./app/locales/en')
-}
-
-export default {
-  messagesByLocale
-}
