--- layout.js
@@ -1,2 +1,0 @@
-import Layout from './app/containers/Layout'
-export default Layout
