--- root.js
@@ -1,8 +1,0 @@
-const appNode = document.createElement('div')
-appNode.id = 'app'
-document.body.appendChild(appNode)
-
-export default {
-  appNode: '#app',
-  styleNode: '#app-styles',
-}
