--- store.js
@@ -1,7 +1,0 @@
-import updater from './updater'
-import epic from './epic'
-
-export default {
-  updater,
-  epic
-}
